DPSPS-431: MVP implementation of Gov Notify support
If GovNotify is configured we fetch all the templates from the GovNotify
account.

If a template has a name which matches the email.type field in the email
table, we disregard the text/html attached to the email and we pull the
template out of gov notify.

We interogate the template parameters, provide them and send the email
via gov notify. We populate the template with only the data it asks for.
This is because some emails ask for passportNumber and we neither want
to send it to all templates, or put email specific rules into this
generic and open source app. We interogate template parameters as a
way of reducing the amount of business logic/solution specific code
here, as well as giving more flexability to the business. Since the
business can interogate the JSON in our app via Kibana or Mongo they
can explore and add template fields as they need, without developer
assistance.

Some examples of what you can use in a GovNotify template include:

((case:registeredTravellerNumber))
((config:selfServeUrl))
((case:currentDetails.passportDetails.passportNumber:right4))
((parent.currentDetails.passportDetails.givenName))

as long it matches what's available in your Mongo document it should
work. If a value is missing, a warning about the template is logged
but it continues with an empty string.

Splitting the template field on : breaks it into three pieces:

* the place to find the value: the email's associated case object
  (prefixed with case:) the config object or parent case object.

* the field. These map one-to-one with Kibana/Mongo so business users
  can find them.

* a function (or list of them) to apply at the end. Example functions
  include :bool (which turns a boolean into GovNotify's yes/no format),
  :not which reverses it so you can replicate our legacy template's
  if/else syntax in GovNotify with

  (case:under18:bool?? You're under 18!)
  (case:under18:bool:not?? You're over 18!)

Tests for this code is required.