DPSPS-1562: Abandoned work to add attachment support to rt-emailer
The organisation want to put PDF receipts on some emails to reduce some
of the corrospondence we get. There are couple of factors which make this
work "not desirable".

* Mongo keeps data as JSON so storing the attachment as a binary format
means keeping base64 data in the email table, which causes is to grow
enourmously.

* Deleting attachments after sending them would be a problem for resend
functionality

* Our email table in some systems is over 20m records. Unfortunately some
apps work by search the email table for reminder emails, or expiry emails to
work out if something has been sent before, so our email tables are destined
to grow continously for a long time.

* We don't want to put app specific business logic and templates into rt-emailer
which is already used in production across three separate systems

* The right place to store binary data in Mongo is via gridfs, not by putting
it in the email table. However rt-emailer also supports Postgres as a backend so
we'd need a second solution here. It further unbalances the feature set for the apps.

* We have recently (in the last week) rewritten all the Mongo driver logic and ported
our libraries such as this one to Scala 2.13. Many of the changes are not fullly deployed
yet and trying to edit upstream libraries and do the switch over is adding complexity.

As a result, we're parking this ticket and logic for now

--

This changeset is a minor proof of concept that shouldn't really be deployed. It adds
attachment:receipt.pdf template keywords and uses that to take binary data from a data
structure in the email table here:

  email.personalisations.attachments: [ "filename" : .., "base64data" : ... }

and uses that to create an attachment in gov notify. This was done to prototype what an
email would look like.

In any case, it's parked for now, or at least until after the Scala 2.13 migration is
complete.